feat: job-group/links api

+++ src/main/kotlin/apply/application/PageDto.kt
@@ -5,7 +5,7 @@ import java.time.LocalDateTime
 data class Page<T> (
     val page_no: Int,
     val page_size: Int,
-    val total_count: Int,
+    val total_count: Long,
     val total_page: Int,
     val contents: List<T>
 )
